# src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IAvaliacao": {
				"properties": {
					"idAvaliacao": {
						"type": "number",
						"format": "double"
					},
					"idCliente": {
						"type": "number",
						"format": "double"
					},
					"idEstabelecimento": {
						"type": "number",
						"format": "double"
					},
					"quantidadeReservas": {
						"type": "number",
						"format": "double"
					},
					"quantidadeCancelamentos": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idAvaliacao",
					"idCliente",
					"idEstabelecimento",
					"quantidadeReservas",
					"quantidadeCancelamentos",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICliente": {
				"properties": {
					"idCliente": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"telefone": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idCliente",
					"nome",
					"email",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEndereco": {
				"properties": {
					"idEndereco": {
						"type": "number",
						"format": "double"
					},
					"endereco": {
						"type": "string"
					},
					"numero": {
						"type": "string"
					},
					"bairro": {
						"type": "string"
					},
					"cidade": {
						"type": "string"
					},
					"uf": {
						"type": "string"
					},
					"pais": {
						"type": "string",
						"format": "date-time"
					},
					"cep": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idEndereco",
					"endereco",
					"numero",
					"bairro",
					"cidade",
					"uf",
					"pais",
					"cep",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEstabelecimento": {
				"properties": {
					"idEstabelecimento": {
						"type": "number",
						"format": "double"
					},
					"idEndereco": {
						"type": "number",
						"format": "double"
					},
					"endereco": {
						"$ref": "#/components/schemas/IEndereco"
					},
					"cnpj": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					},
					"telefone": {
						"type": "string"
					},
					"lotacao": {
						"type": "string"
					},
					"horaAbertura": {
						"type": "string",
						"format": "date-time"
					},
					"descricao": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idEstabelecimento",
					"idEndereco",
					"endereco",
					"cnpj",
					"senha",
					"telefone",
					"lotacao",
					"horaAbertura",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFidelidade": {
				"properties": {
					"idFidelidade": {
						"type": "number",
						"format": "double"
					},
					"idCliente": {
						"type": "number",
						"format": "double"
					},
					"idEstabelecimento": {
						"type": "number",
						"format": "double"
					},
					"quantidadeReservas": {
						"type": "number",
						"format": "double"
					},
					"quantidadeCancelamentos": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idFidelidade",
					"idCliente",
					"idEstabelecimento",
					"quantidadeReservas",
					"quantidadeCancelamentos",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IReserva": {
				"properties": {
					"idReserva": {
						"type": "number",
						"format": "double"
					},
					"idCliente": {
						"type": "number",
						"format": "double"
					},
					"idEstabelecimento": {
						"type": "number",
						"format": "double"
					},
					"valor": {
						"type": "string"
					},
					"dataReserva": {
						"type": "string",
						"format": "date-time"
					},
					"horaInicio": {
						"type": "string",
						"format": "date-time"
					},
					"horaFim": {
						"type": "string",
						"format": "date-time"
					},
					"qtdPessoas": {
						"type": "string"
					}
				},
				"required": [
					"idReserva",
					"idCliente",
					"idEstabelecimento",
					"valor",
					"dataReserva",
					"horaInicio",
					"horaFim",
					"qtdPessoas"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend-covid-19-reservation-app",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/avaliacoes/{estabelecimentoID}": {
			"get": {
				"operationId": "GetAvaliacaoEstabelecimento",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IAvaliacao"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Avaliacoes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "estabelecimentoID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/avaliacoes": {
			"post": {
				"operationId": "CriarAvaliacao",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAvaliacao"
								}
							}
						}
					}
				},
				"tags": [
					"Avaliacoes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAvaliacao"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarAvaliacao",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAvaliacao"
								}
							}
						}
					}
				},
				"tags": [
					"Avaliacoes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAvaliacao"
							}
						}
					}
				}
			}
		},
		"/avaliacoes/{avaliacaoID}": {
			"delete": {
				"operationId": "DeletarAvaliacao",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Avaliacoes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "avaliacaoID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/clientes": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICliente"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CriarCliente",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICliente"
								}
							}
						}
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICliente"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarClietne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICliente"
								}
							}
						}
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICliente"
							}
						}
					}
				}
			}
		},
		"/clientes/{clienteID}": {
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clienteID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/enderecos": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IEndereco"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Enderecos"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CriarEndereco",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEndereco"
								}
							}
						}
					}
				},
				"tags": [
					"Enderecos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEndereco"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarEndereco",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEndereco"
								}
							}
						}
					}
				},
				"tags": [
					"Enderecos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEndereco"
							}
						}
					}
				}
			}
		},
		"/enderecos/{enderecoID}": {
			"delete": {
				"operationId": "DeletarCliente",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Enderecos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "enderecoID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/estabelecimentos": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IEstabelecimento"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Estabelecimentos"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cidade",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "uf",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CriarEstabelecimento",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEstabelecimento"
								}
							}
						}
					}
				},
				"tags": [
					"Estabelecimentos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEstabelecimento"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarEstabelecimento",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEstabelecimento"
								}
							}
						}
					}
				},
				"tags": [
					"Estabelecimentos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEstabelecimento"
							}
						}
					}
				}
			}
		},
		"/estabelecimentos/{estabelecimentoID}": {
			"delete": {
				"operationId": "DeletarEstabelecimento",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Estabelecimentos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "estabelecimentoID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/fidelidade": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IFidelidade"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Fidelidade"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "idEstabelecimento",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "idCliente",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AtualizarFidelidade",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFidelidade"
								}
							}
						}
					}
				},
				"tags": [
					"Fidelidade"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IFidelidade"
							}
						}
					}
				}
			}
		},
		"/fidelidade/{FidelidadeID}": {
			"delete": {
				"operationId": "DeletarFidelidade",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Fidelidade"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "FidelidadeID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reservas/cliente/{clienteID}": {
			"get": {
				"operationId": "GetReservasByClienteID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IReserva"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clienteID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dataPesquisa",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/reservas/estabelecimento/{estabelecimentoID}": {
			"get": {
				"operationId": "GetReservasByEstabelecimentoID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IReserva"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "estabelecimentoID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dataPesquisa",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/reservas": {
			"post": {
				"operationId": "CriarReserva",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReserva"
								}
							}
						}
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReserva"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarReserva",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReserva"
								}
							}
						}
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReserva"
							}
						}
					}
				}
			}
		},
		"/reservas/{reservaID}": {
			"delete": {
				"operationId": "DeletarReserva",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "reservaID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}